chore: add crypto callbacks to src
